Move all model directory functions to a single `utils.py` file
